==> 02-feature-engineering.ipynb <==
# Feature Engineering

import numpy as np
import pandas as pd

df = pd.read_parquet("solar_cleaned.parquet")

middle = int(len(df) * 0.7)

train = df.loc[:middle].copy()
test = df.loc[middle:]
print("Training shape: " + str(train.shape))
print("Testing shape: " + str(test.shape))

Convert hourly data into aggrigated daily data. For categorized columns, split each column up. the rest take an averge, min and max. 

import altair as alt

station = daily_summary.loc[171,:]

source = pd.DataFrame({
    'dni_efficiency': station['dni_efficiency'],
    'cloud_cover': station['cloud_cover'],
    'temperature': station['temperature'],

})


base = alt.Chart(source).mark_circle(opacity=0.5).transform_fold(
    fold=['cloud_cover'],
    as_=['category', 'y']
).encode(
    alt.X('dni_efficiency:Q'),
    alt.Y('y:Q'),
    alt.Color('category:N')
)

base + base.transform_loess('dni_efficiency', 'y', groupby=['category']).mark_line(size=4)


drop columns that will not have an impact

from sklearn.preprocessing import RobustScaler
from sklearn.impute import SimpleImputer
from sklearn.pipeline import Pipeline
from sklearn.compose import ColumnTransformer

columns_to_scale = list(test.drop(columns=['dni_efficiency', 'ghi_efficiency', 'dhi_efficiency', 'STATION','DATE',
                                            'latitude','longitude']).columns)

amount_scaler = ('amount_scaler', RobustScaler(), columns_to_scale)

scale_steps = [amount_scaler]
all_xforms = ColumnTransformer(transformers=(scale_steps))

feat_pipeline = Pipeline([
    ('feature_extraction',all_xforms)
])

feat_pipeline.fit(train)

from mlworkflows import util
util.serialize_to(feat_pipeline, "feature_pipeline.sav")
